Use builting replaceAll everywhere
This allows for two helper functions to be removed and for the code to
be clearer and more efficient. The string is iterated through fewer
times as the number of replacements does not need to be found before
performing the replacement. Previously this was atleast two linear
searches for each replacement.